+++ docs/changes.rst
@@ -25,6 +25,11 @@ Released: not yet
 
 **Bug fixes:**
 
+* Fixed the issue that INSTANCE child elements on a returned ERROR element
+  were not allowed. INSTANCE child elements are now allowed and will appear
+  to the user as a list of `CIMInstance` objects in a new `instances` property
+  of the `CIMError` exception that is raised. See issue #1380.
+
 **Enhancements:**
 
 * Make: Eliminated the confusing but unproblematic error message about
@@ -44,6 +49,10 @@ Released: not yet
   internal attributes and no longer the string representation of the value.
   Added a `__repr__()` method to the `pywbem.MinutesFromUTC` class that shows
   its internal attributes. See issue #1379.
+
+* Added an `instances` property to the `CIMError` exception class that can
+  be used to represent a list of error instances returned by the WBEM server
+  in error responses. See issue #1380.
 
 **Known issues:**